move out TabLabel
