+++ plugins/parodos/src/components/ParodosPage/index.ts
@@ -1,0 +1,1 @@
+export { ParodosPage } from './ParodosPage';
